pyproject.toml
[project]
name = "intro-to-ai"
version = "0.1.0"
description = "Introduction to AI: Practical Python implementations of core AI concepts"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]

dependencies = [
    # Core scientific computing
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "pandas>=2.1.0",

    # Machine Learning
    "scikit-learn>=1.3.0",

    # Deep Learning
    "tensorflow>=2.15.0",
    "keras>=3.0.0",

    # NLP
    "nltk>=3.8.0",
    "spacy>=3.7.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",

    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",

    # Interactive UI
    "gradio>=4.0.0",

    # Graph algorithms
    "networkx>=3.2.0",

    # Probabilistic programming
    "pgmpy>=0.1.25",

    # Symbolic math
    "sympy>=1.12",

    # Computer vision
    "opencv-python>=4.8.0",
    "pillow>=10.1.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",

    # Jupyter
    "jupyter>=1.0.0",
    "ipywidgets>=8.1.0",
    "ipykernel>=6.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
